Sync the TrashCounter, use PredictedSpawn
- Spawning and despawning kitchen objects is now predicted and happens
instantly for the client. While the actual spawning and despawning is
handled by Fishnet, there is some additional logic for animations etc.
that assumes the spawns have worked on the server. This could break the
game if there some packet loss causing the spawn to not actually go
through. Need to figure out the canonical way of rolling back logic that
relies on a successful predicted spawn in the case that it fails.