Adds a ManagersLoadingScene
Players now move from the lobby to the empty ManagersLoadingScene, and
once the NetworkConnections script has loaded on the client, they move
to the character select screen. This guarantees that the
NetworkConnections script will be available on the client to all scripts
in subsequent scenes as soon as they initiate.

The LobbyPlayerConnection object is no longer persisted across scenes
and just triggers events to be used by the lobby UI on connection and
rejection.