Fixed matchmaker rate limit bug
- bug was caused by the PollMatchmakerTicket method sending too many
GetTicketAsync requests simultaneously. Solution was to set a flag to
true when the method started and false when it completed. And then only
countdown the timer to the next call to PollMatchmakerTicket when the
flag was false. So only after the previous call completed would it then
wait 1.1 seconds before calling again.